Split Prices and Yields for Securities
