+++ src/PhpSpreadsheet/Calculation/Financial/Securities/Constants.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheet\Calculation\Financial\Securities;
+
+class Constants
+{
+    public const FREQUENCY_ANNUAL = 1;
+    public const FREQUENCY_SEMI_ANNUAL = 2;
+    public const FREQUENCY_QUARTERLY = 4;
+}
